# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ixnos-data API",
    "description": "Search Greek public procurement and spending data from ΚΗΜΔΗΣ and Διαύγεια.\n\n**Rate limits.** Without a key, 120 requests a minute per IP address. With an API key\n(free: sign in on the website, then create one on the alerts page), send it as the\n`X-Api-Key` header for 600 requests a minute. Over the limit: HTTP 429.\n\n**Bulk data.** Daily exports of every record (CSV and JSON Lines) are at /exports/.\n\n**Sources and licence.** ΚΗΜΔΗΣ and Διαύγεια open data, both CC BY 4.0: credit the\nsources when you reuse the data.",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "v1"
  },
  "paths": {
    "/v1/search": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Search records",
        "description": "Full-text search in Greek (accents and case ignored, words stemmed), typo-tolerant, Greeklish-aware, with filters. An ΑΔΑΜ or ΑΔΑ finds that record; a nine-digit VAT number finds records of that organisation or contractor.",
        "operationId": "SearchItems",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "cpv",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nuts",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organisation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "maxAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signal",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchItemsResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{sourceId}": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Get one record by its ΑΔΑΜ (ΚΗΜΔΗΣ) or ΑΔΑ (Διαύγεια)",
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/organisations/{id}": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Get a contracting authority with its activity and recent records",
        "operationId": "GetOrganisation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/cpv": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Find CPV codes by label (Greek or English) or code prefix",
        "description": "For picking a trade: at most 12 codes, broad categories first. Needs at least 2 characters.",
        "operationId": "FindCpv",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodeLabel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sitemap/items/{page}": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Record identifiers for search engines, 50000 per page",
        "operationId": "GetSitemapItems",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SitemapEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeLabel": {
        "required": [
          "code",
          "labelEl",
          "labelEn"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "labelEl": {
            "type": [
              "null",
              "string"
            ]
          },
          "labelEn": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ContractorSpend": {
        "required": [
          "name",
          "awards",
          "amountEur",
          "paidEur"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "awards": {
            "type": "integer",
            "format": "int32"
          },
          "amountEur": {
            "type": "number",
            "format": "double"
          },
          "paidEur": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "DominantSupplier": {
        "required": [
          "name",
          "share",
          "awards",
          "totalAwards"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "share": {
            "type": "number",
            "format": "double"
          },
          "awards": {
            "type": "integer",
            "format": "int32"
          },
          "totalAwards": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ItemContractorView": {
        "required": [
          "name",
          "role",
          "amountEur",
          "countryCode"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "amountEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "countryCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ItemDetail": {
        "required": [
          "sourceId",
          "source",
          "kind",
          "title",
          "description",
          "organisation",
          "amountEur",
          "amountWithVatEur",
          "publishedAt",
          "signedOn",
          "deadlineAt",
          "startsOn",
          "endsOn",
          "cancelled",
          "cancelledOn",
          "cpv",
          "nuts",
          "contractors",
          "links",
          "documentUrl",
          "supersededBy"
        ],
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "organisation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrganisationRef"
              }
            ]
          },
          "amountEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "amountWithVatEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "signedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "deadlineAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "startsOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "endsOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "cancelled": {
            "type": "boolean"
          },
          "cancelledOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "cpv": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeLabel"
            }
          },
          "nuts": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CodeLabel"
              }
            ]
          },
          "contractors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemContractorView"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemLinkView"
            }
          },
          "documentUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "supersededBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "procedure": {
            "type": [
              "null",
              "string"
            ]
          },
          "offersReceived": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "signals": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "amountImplausible": {
            "type": "boolean"
          }
        }
      },
      "ItemLinkView": {
        "required": [
          "relation",
          "source",
          "sourceId",
          "kind",
          "title"
        ],
        "type": "object",
        "properties": {
          "relation": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ItemSummary": {
        "required": [
          "sourceId",
          "source",
          "kind",
          "title",
          "organisation",
          "amountEur",
          "publishedAt",
          "deadlineAt",
          "nutsCode",
          "cpvCodes",
          "cancelled",
          "amountWithVatEur"
        ],
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "organisation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrganisationRef"
              }
            ]
          },
          "amountEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deadlineAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "nutsCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "cpvCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cancelled": {
            "type": "boolean"
          },
          "amountWithVatEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "signals": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "amountImplausible": {
            "type": "boolean"
          }
        }
      },
...
      "OrganisationDetail": {
        "required": [
          "id",
          "nameEl",
          "nameEn",
          "type",
          "taxId",
          "parent",
          "website",
          "itemCounts",
          "awardedLast12MonthsEur",
          "paidLast12MonthsEur",
          "awardedByYear",
          "topContractors",
          "recentItems"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nameEl": {
            "type": "string"
          },
          "nameEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "taxId": {
            "type": [
              "null",
              "string"
            ]
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrganisationRef"
              }
            ]
          },
          "website": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "itemCounts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "awardedLast12MonthsEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "paidLast12MonthsEur": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "awardedByYear": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/YearSpend"
            }
          },
          "topContractors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContractorSpend"
            }
          },
          "recentItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            }
          },
          "dominantSupplier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DominantSupplier"
              }
            ]
          }
        }
      },
      "OrganisationRef": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "SearchItemsResult": {
        "required": [
          "total",
          "page",
          "pageSize",
          "mode",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "mode": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            }
          },
          "totalCapped": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "SitemapEntry": {
        "required": [
          "id",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "YearSpend": {
        "required": [
          "year",
          "awards",
          "amountEur",
          "approvedEur",
          "paidEur"
        ],
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "awards": {
            "type": "integer",
            "format": "int32"
          },
          "amountEur": {
            "type": "number",
            "format": "double"
          },
          "approvedEur": {
            "type": "number",
            "format": "double"
          },
          "paidEur": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  },
...
}